+++ Skyflow/src/main/kotlin/Skyflow/CollectOptions.kt
@@ -1,6 +1,7 @@
 package Skyflow
 
+import org.json.JSONArray
 import org.json.JSONObject
 
-class CollectOptions(val token:Boolean = true, val additionalFields: JSONObject? = null) {
+class CollectOptions(val token:Boolean = true, val additionalFields: JSONObject? = null, val upsert : JSONArray? = null) {
 }